Added PLC Hash to the beginning of modbus map
The beginning (first 10 words) of the modbus map is stable, meaning that it
does not depend on the size of the data stream. The problem with the S7 stream
is that we don't know where it begins; we rely on knowing the size of
the data stream. If we don't know the data stream we cannot tell what
Hash the PLC is using.