Support 'gateway-mode' with Siemens PLCs
Safety PLCs have a gateway PLC to shield them from the world. This
gateway PLC has a giant datablock containing all the PLC variables from
the safety PLC that need to be exported to EPICS. 'Gateway-mode' reads
the variables dirctly from this datablock instead of duplicating them in
instance DBs. Specify the name of the gateway datablock in the
'PLC-EPICS-COMMS: GatewayDatablock' property of the (gateway) PLC